refactor: reuse prerender util
